Move ReactiveTransactionsWrapper tests into a new file
