+++ src/test/java/org/springframework/data/couchbase/transactions/CouchbasePersonTransactionIntegrationTests.java
@@ -71,7 +71,6 @@ import org.springframework.data.couchbase.transaction.CouchbaseSimpleCallbackTra
 import org.springframework.data.couchbase.transaction.CouchbaseTransactionDefinition;
 import org.springframework.data.couchbase.transaction.ReactiveCouchbaseResourceHolder;
 import org.springframework.data.couchbase.transaction.ReactiveCouchbaseTransactionManager;
-import org.springframework.data.couchbase.transaction.ReactiveTransactionsWrapper;
 import org.springframework.data.couchbase.transaction.TransactionsWrapper;
 import org.springframework.data.couchbase.util.Capabilities;
 import org.springframework.data.couchbase.util.ClusterType;
@@ -127,7 +126,6 @@ public class CouchbasePersonTransactionIntegrationTests extends JavaIntegrationT
 	String sName = "_default";
 	String cName = "_default";
 	private TransactionalOperator transactionalOperator;
-	private ReactiveTransactionsWrapper reactiveTransactionsWrapper;
 
 	@BeforeAll
 	public static void beforeAll() {
@@ -158,8 +156,6 @@ public class CouchbasePersonTransactionIntegrationTests extends JavaIntegrationT
 		Person walterWhite = new Person(1, "Walter", "White");
 		remove(cbTmpl, sName, cName, walterWhite.getId().toString());
 		transactionalOperator = TransactionalOperator.create(reactiveCouchbaseTransactionManager);
-		reactiveTransactionsWrapper = new ReactiveTransactionsWrapper(
-				reactiveCouchbaseClientFactory);
 	}
 
 	@Test
@@ -335,32 +331,6 @@ public class CouchbasePersonTransactionIntegrationTests extends JavaIntegrationT
 		assertEquals(switchedPerson.getFirstname(), pFound.getFirstname(), "should have been switched");
 	}
 
-	@Test
-	public void wrapperReplaceWithCasConflictResolvedViaRetryReactive() {
-		Person person = new Person(1, "Walter", "White");
-		Person switchedPerson = new Person(1, "Dave", "Reynolds");
-		AtomicInteger tryCount = new AtomicInteger(0);
-		cbTmpl.insertById(Person.class).one(person);
-
-		for (int i = 0; i < 50; i++) { // the transaction sometimes succeeds on the first try
-			ReplaceLoopThread t = new ReplaceLoopThread(switchedPerson);
-			t.start();
-			tryCount.set(0);
-			Mono<TransactionResult> result = reactiveTransactionsWrapper.run(ctx -> {
-				System.err.println("try: " + tryCount.incrementAndGet());
-				return rxCBTmpl.findById(Person.class).one(person.getId().toString()) //
-						.flatMap(ppp -> rxCBTmpl.replaceById(Person.class).one(ppp)).then();
-			});
-			TransactionResult txResult = result.block();
-			t.setStopFlag();
-			if (tryCount.get() > 1) {
-				break;
-			}
-		}
-		Person pFound = cbTmpl.findById(Person.class).one(person.getId().toString());
-		assertTrue(tryCount.get() > 1, "should have been more than one try. tries: " + tryCount.get());
-		assertEquals(switchedPerson.getFirstname(), pFound.getFirstname(), "should have been switched");
-	}
 
 	/**
 	 * This does process retries - by CallbackTransactionManager.execute() -> transactions.run() -> executeTransaction()
@@ -444,137 +414,6 @@ public class CouchbasePersonTransactionIntegrationTests extends JavaIntegrationT
 		System.out.println("pFound: " + pFound);
 		assertEquals(switchedPerson.getFirstname(), pFound.getFirstname(), "should have been switched");
 		assertTrue(tryCount.get() > 1, "should have been more than one try. tries: " + tryCount.get());
-	}
-
-	@Test
-	public void replacePersonCBTransactionsRxTmplRollback() {
-		Person person = new Person(1, "Walter", "White");
-		String newName = "Walt";
-		cbTmpl.insertById(Person.class).one(person);
-		Mono<TransactionResult> result = reactiveTransactionsWrapper.run(ctx -> { //
-			return rxCBTmpl.findById(Person.class).one(person.getId().toString()) //
-					.flatMap(pp -> rxCBTmpl.replaceById(Person.class).one(pp.withFirstName(newName))).then(Mono.empty());
-		});
-		result.block();
-		Person pFound = cbTmpl.findById(Person.class).one(person.getId().toString());
-		System.err.println(pFound);
-		assertEquals(newName, pFound.getFirstname());
-	}
-
-	@Test
-	public void deletePersonCBTransactionsRxTmpl() {
-		Person person = new Person(1, "Walter", "White");
-		remove(cbTmpl, sName, cName, person.getId().toString());
-		cbTmpl.insertById(Person.class).inCollection(cName).one(person);
-		Mono<TransactionResult> result = reactiveTransactionsWrapper.run(ctx -> { // get the ctx
-			return rxCBTmpl.removeById(Person.class).inCollection(cName).oneEntity(person).then();
-		});
-		result.block();
-		Person pFound = cbTmpl.findById(Person.class).inCollection(cName).one(person.getId().toString());
-		assertNull(pFound, "Should not have found " + pFound);
-	}
-
-	@Test
-	public void deletePersonCBTransactionsRxTmplFail() {
-		Person person = new Person(1, "Walter", "White");
-		remove(cbTmpl, sName, cName, person.getId().toString());
-		cbTmpl.insertById(Person.class).inCollection(cName).one(person);
-		Mono<TransactionResult> result = reactiveTransactionsWrapper.run(ctx -> { // get the ctx
-			return rxCBTmpl.removeById(Person.class).inCollection(cName).oneEntity(person)
-					.then(rxCBTmpl.removeById(Person.class).inCollection(cName).oneEntity(person)).then();
-		});
-		assertThrowsWithCause(result::block, TransactionFailedException.class, DataRetrievalFailureException.class);
-		Person pFound = cbTmpl.findById(Person.class).inCollection(cName).one(person.getId().toString());
-		assertEquals(pFound, person, "Should have found " + person);
-	}
-
-	@Test
-	public void deletePersonCBTransactionsRxRepo() {
-		Person person = new Person(1, "Walter", "White");
-		remove(cbTmpl, sName, cName, person.getId().toString());
-		repo.withCollection(cName).save(person);
-		Mono<TransactionResult> result = reactiveTransactionsWrapper.run(ctx -> { // get the ctx
-			return rxRepo.withCollection(cName).delete(person).then();
-		});
-		result.block();
-		Person pFound = cbTmpl.findById(Person.class).inCollection(cName).one(person.getId().toString());
-		assertNull(pFound, "Should not have found " + pFound);
-	}
-
-	@Test
-	public void deletePersonCBTransactionsRxRepoFail() {
-		Person person = new Person(1, "Walter", "White");
-		remove(cbTmpl, sName, cName, person.getId().toString());
-		repo.withCollection(cName).save(person);
-		Mono<TransactionResult> result = reactiveTransactionsWrapper.run(ctx -> { // get the ctx
-			return rxRepo.withCollection(cName).findById(person.getId().toString())
-					.flatMap(pp -> rxRepo.withCollection(cName).delete(pp).then(rxRepo.withCollection(cName).delete(pp))).then();
-		});
-		assertThrowsWithCause(result::block, TransactionFailedException.class, DataRetrievalFailureException.class);
-		Person pFound = cbTmpl.findById(Person.class).inCollection(cName).one(person.getId().toString());
-		assertEquals(pFound, person, "Should have found " + person + " instead of " + pFound);
-	}
-
-	@Test
-	public void findPersonCBTransactions() {
-		Person person = new Person(1, "Walter", "White");
-		remove(cbTmpl, sName, cName, person.getId().toString());
-		cbTmpl.insertById(Person.class).inScope(sName).inCollection(cName).one(person);
-		List<Object> docs = new LinkedList<>();
-		Query q = Query.query(QueryCriteria.where("meta().id").eq(person.getId()));
-		Mono<TransactionResult> result = reactiveTransactionsWrapper.run(ctx -> rxCBTmpl.findByQuery(Person.class)
-				.inScope(sName).inCollection(cName).matching(q).withConsistency(REQUEST_PLUS).one().doOnSuccess(doc -> {
-					System.err.println("doc: " + doc);
-					docs.add(doc);
-				}));
-		result.block();
-		assertFalse(docs.isEmpty(), "Should have found " + person);
-		for (Object o : docs) {
-			assertEquals(o, person, "Should have found " + person + " instead of " + o);
-		}
-	}
-
-	@Test
-	public void insertPersonRbCBTransactions() {
-		Person person = new Person(1, "Walter", "White");
-		remove(cbTmpl, sName, cName, person.getId().toString());
-		Mono<TransactionResult> result = reactiveTransactionsWrapper
-				.run(ctx -> rxCBTmpl.insertById(Person.class).inScope(sName).inCollection(cName).one(person)
-						.<Person> flatMap(it -> Mono.error(new SimulateFailureException())));
-		assertThrowsWithCause(() -> result.block(), TransactionFailedException.class, SimulateFailureException.class);
-		Person pFound = cbTmpl.findById(Person.class).inCollection(cName).one(person.getId().toString());
-		assertNull(pFound, "Should not have found " + pFound);
-	}
-
-	@Test
-	public void replacePersonRbCBTransactions() {
-		Person person = new Person(1, "Walter", "White");
-		remove(cbTmpl, sName, cName, person.getId().toString());
-		cbTmpl.insertById(Person.class).inScope(sName).inCollection(cName).one(person);
-		Mono<TransactionResult> result = reactiveTransactionsWrapper.run(ctx -> //
-		rxCBTmpl.findById(Person.class).inScope(sName).inCollection(cName).one(person.getId().toString()) //
-				.flatMap(pFound -> rxCBTmpl.replaceById(Person.class).inScope(sName).inCollection(cName)
-						.one(pFound.withFirstName("Walt")))
-				.<Person> flatMap(it -> Mono.error(new SimulateFailureException())));
-		assertThrowsWithCause(() -> result.block(), TransactionFailedException.class, SimulateFailureException.class);
-		Person pFound = cbTmpl.findById(Person.class).inScope(sName).inCollection(cName).one(person.getId().toString());
-		assertEquals(person, pFound, "Should have found " + person + " instead of " + pFound);
-	}
-
-	@Test
-	public void findPersonSpringTransactions() {
-		Person person = new Person(1, "Walter", "White");
-		remove(cbTmpl, sName, cName, person.getId().toString());
-		cbTmpl.insertById(Person.class).inScope(sName).inCollection(cName).one(person);
-		List<Object> docs = new LinkedList<>();
-		Query q = Query.query(QueryCriteria.where("meta().id").eq(person.getId()));
-		Mono<TransactionResult> result = reactiveTransactionsWrapper.run(ctx -> rxCBTmpl.findByQuery(Person.class)
-				.inScope(sName).inCollection(cName).matching(q).one().doOnSuccess(r -> docs.add(r)));
-		result.block();
-		assertFalse(docs.isEmpty(), "Should have found " + person);
-		for (Object o : docs) {
-			assertEquals(o, person, "Should have found " + person);
-		}
 	}
 
 	private class ReplaceLoopThread extends Thread {
